Initial code commit for tuition repayment plans is added in this commits. I also added tests.
Astute observers will notice this addition is very different from the way the
existing code was written. This is a kind of experiment, which I think I like.
Given time, I'll rehabilitate this code so that is consists of more modules,
each of which contains methods which take in the 'data' Object and return a
single value, then string those together. It ought to make the code more
readable.