More code for tuition repayment plans.
Due to the very odd nature of these financial vehicles, I've decided NOT to
combine the debt or monthly payments with the rest of the loan payments. These
values are still passed in the Object returned by the package, so adding the
tuition repayment monthly payment to the total loan monthly payment is easy
enough in the application using this package.

This made me think - version 3 of this package might be better if multiple
individual totals were passed so that the application developer can choose
what values they need. It's worth investigating.